Complete Overall (see full commit message)
Previously, byte channels were used for the greater part of the
communication. This approach was somewhat inflexible, one big problem
being that getchar() always returned 3 bytes for one key press of an
arrow key (these keys are used for many games).

The new approach is based on the 'Key' struct in the 'net' package.
It just wraps an 'goncurses.Key', but we need a struct because the data
is now transferred via 'encoding/gob'.

Almost all occurences of byte channel are now net.Key channels.

The new implementation should be a better foundation for multi OS
support and extensibility.

It's still buggy though...